Remove TaskViewFactory.kt and move the helper method to DataCollectionViewModel.kt
