--- ground/src/main/java/com/google/android/ground/ui/editsubmission/TaskViewFactory.kt
@@ -1,52 +1,0 @@
-/*
- * Copyright 2020 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.google.android.ground.ui.editsubmission
-
-import androidx.fragment.app.Fragment
-import com.google.android.ground.model.task.Task
-import com.google.android.ground.ui.common.ViewModelFactory
-import com.google.android.ground.ui.datacollection.AbstractTaskViewModel
-import com.google.android.ground.ui.datacollection.tasks.date.DateTaskViewModel
-import com.google.android.ground.ui.datacollection.tasks.multiplechoice.MultipleChoiceTaskViewModel
-import com.google.android.ground.ui.datacollection.tasks.number.NumberTaskViewModel
-import com.google.android.ground.ui.datacollection.tasks.photo.PhotoTaskViewModel
-import com.google.android.ground.ui.datacollection.tasks.point.DropAPinTaskViewModel
-import com.google.android.ground.ui.datacollection.tasks.polygon.PolygonDrawingViewModel
-import com.google.android.ground.ui.datacollection.tasks.text.TextTaskViewModel
-import com.google.android.ground.ui.datacollection.tasks.time.TimeTaskViewModel
-import javax.inject.Inject
-
-/** Inflates a new view and generates a view model for a given [Task.Type]. */
-class TaskViewFactory @Inject internal constructor() {
-
-  @Inject lateinit var fragment: Fragment
-  @Inject lateinit var viewModelFactory: ViewModelFactory
-
-  companion object {
-    fun getViewModelClass(taskType: Task.Type): Class<out AbstractTaskViewModel> =
-      when (taskType) {
-        Task.Type.TEXT -> TextTaskViewModel::class.java
-        Task.Type.MULTIPLE_CHOICE -> MultipleChoiceTaskViewModel::class.java
-        Task.Type.PHOTO -> PhotoTaskViewModel::class.java
-        Task.Type.NUMBER -> NumberTaskViewModel::class.java
-        Task.Type.DATE -> DateTaskViewModel::class.java
-        Task.Type.TIME -> TimeTaskViewModel::class.java
-        Task.Type.DROP_A_PIN -> DropAPinTaskViewModel::class.java
-        Task.Type.DRAW_POLYGON -> PolygonDrawingViewModel::class.java
-        Task.Type.UNKNOWN -> throw IllegalArgumentException("Unsupported task type: $taskType")
-      }
-  }
-}
